+++ ibtk/src/utilities/ParallelMap.cpp
@@ -14,8 +14,8 @@
 //      notice, this list of conditions and the following disclaimer in the
 //      documentation and/or other materials provided with the distribution.
 //
-//    * Neither the name of New York University nor the names of its
-//      contributors may be used to endorse or promote products derived from
+//    * Neither the name of The University of North Carolina nor the names of
+//      its contributors may be used to endorse or promote products derived from
 //      this software without specific prior written permission.
 //
 // THIS SOFTWARE IS PROVIDED BY THE COPYRIGHT HOLDERS AND CONTRIBUTORS "AS IS"
@@ -32,17 +32,19 @@
 
 /////////////////////////////// INCLUDES /////////////////////////////////////
 
+#include <map>
 #include <ostream>
 #include <utility>
 #include <vector>
 
 #include "IntVector.h"
-#include "ParallelMap.h"
 #include "ibtk/FixedSizedStream.h"
+#include "ibtk/ParallelMap.h"
 #include "ibtk/Streamable.h"
 #include "ibtk/StreamableManager.h"
 #include "ibtk/namespaces.h" // IWYU pragma: keep
 #include "tbox/AbstractStream.h"
+#include "tbox/Pointer.h"
 #include "tbox/SAMRAI_MPI.h"
 #include "tbox/Utilities.h"
 
@@ -61,8 +63,7 @@ ParallelMap::ParallelMap() : d_map(), d_pending_additions(), d_pending_removals(
 } // ParallelMap
 
 ParallelMap::ParallelMap(const ParallelMap& from)
-    : d_map(from.d_map), d_pending_additions(from.d_pending_additions),
-      d_pending_removals(from.d_pending_removals)
+    : d_map(from.d_map), d_pending_additions(from.d_pending_additions), d_pending_removals(from.d_pending_removals)
 {
     // intentionally blank
     return;
@@ -110,15 +111,14 @@ void ParallelMap::communicateData()
         // Determine how many keys have been registered for addition on each
         // process.
         std::vector<int> num_additions(size, 0);
-        num_additions[rank] = d_pending_additions.size();
+        num_additions[rank] = static_cast<int>(d_pending_additions.size());
         SAMRAI_MPI::sumReduction(&num_additions[0], size);
 
         // Get the local values to send and determine the amount of data to be
         // broadcast by each process.
         std::vector<int> keys_to_send;
         std::vector<tbox::Pointer<Streamable> > data_items_to_send;
-        for (std::map<int, tbox::Pointer<Streamable> >::const_iterator cit =
-                 d_pending_additions.begin();
+        for (std::map<int, tbox::Pointer<Streamable> >::const_iterator cit = d_pending_additions.begin();
              cit != d_pending_additions.end();
              ++cit)
         {
@@ -126,8 +126,8 @@ void ParallelMap::communicateData()
             data_items_to_send.push_back(cit->second);
         }
         std::vector<int> data_sz(size, 0);
-        data_sz[rank] = tbox::AbstractStream::sizeofInt() * keys_to_send.size() +
-                        streamable_manager->getDataStreamSize(data_items_to_send);
+        data_sz[rank] = static_cast<int>(tbox::AbstractStream::sizeofInt() * keys_to_send.size() +
+                                         streamable_manager->getDataStreamSize(data_items_to_send));
         SAMRAI_MPI::sumReduction(&data_sz[0], size);
 
         // Broadcast data from each process.
@@ -139,15 +139,14 @@ void ParallelMap::communicateData()
             {
                 // Pack and broadcast data on process sending_proc.
                 FixedSizedStream stream(data_sz[sending_proc]);
-                stream.pack(&keys_to_send[0], keys_to_send.size());
+                stream.pack(&keys_to_send[0], static_cast<int>(keys_to_send.size()));
                 streamable_manager->packStream(stream, data_items_to_send);
                 int data_size = stream.getCurrentSize();
 #if !defined(NDEBUG)
                 TBOX_ASSERT(static_cast<int>(d_pending_additions.size()) == num_keys);
                 TBOX_ASSERT(data_size == data_sz[sending_proc]);
 #endif
-                SAMRAI_MPI::bcast(
-                    static_cast<char*>(stream.getBufferStart()), data_size, sending_proc);
+                SAMRAI_MPI::bcast(static_cast<char*>(stream.getBufferStart()), data_size, sending_proc);
                 for (int k = 0; k < num_keys; ++k)
                 {
                     d_map[keys_to_send[k]] = data_items_to_send[k];
@@ -188,7 +187,7 @@ void ParallelMap::communicateData()
         // Determine how many keys have been registered for removal on each
         // process.
         std::vector<int> num_removals(size, 0);
-        num_removals[rank] = d_pending_removals.size();
+        num_removals[rank] = static_cast<int>(d_pending_removals.size());
         SAMRAI_MPI::sumReduction(&num_removals[0], size);
 
         // Broadcast data from each process.
